Swimming Warp, Bubble Puppeteer, & more...
I'll explain my favorites.

Swimming Warp is a neat spell wherein you blink and then leap upward. The effect, however, makes it appear as if you swam through the ground and launched upward out of it.

Bubble Puppeteer uses the LevitateSpell, but with the SphereEffect to make it appear as if the target is trapped within a bubble.

Dragon Claw is a really, really cool looking spell using the FireballSpell + the AnimatedBallEffect. I've said it once, and I'll say it again, the AnimatedBallEffect is awesome.