Ever-Filling Soup Bowl, Composer's Baton, Alchemy Bomb
Added three new spells, all in the instant category!

The Ever-Filling Soup Bowl is what you'd expect it to be, you can refill it while standing in water.

The Composer's Baton plays notes from the noteblock (harp). I found the pitches and stuff on the MC wiki (https://minecraft.gamepedia.com/Note_Block) so if it's wrong, yell at them.

The Alchemy Bomb is a projectile which does an AoE effect, applying wither to anything near it.